+++ src/HotChocolate/Core/src/Types/Types/Relay/Extensions/RelaySchemaBuilderExtensions.cs
@@ -57,37 +57,3 @@ public static class RelaySchemaBuilderExtensions
             .TryAddTypeInterceptor<QueryFieldTypeInterceptor>();
     }
 }
-
-// TODO: Move
-public sealed class GlobalObjectIdentificationOptions
-{
-    /// <summary>
-    /// Specifies if the node interface and the node field shall be registered with the schema.
-    /// </summary>
-    public bool RegisterNodeInterface { get; set; } = true;
-
-    /// <summary>
-    /// Defines if the schema building process shall validate that all nodes are resolvable through `node`.
-    /// </summary>
-    public bool EnsureAllNodesCanBeResolved { get; set; } = true;
-
-    /// <summary>
-    /// Specifies the maximum allowed nodes that can be fetched at once through the nodes field.
-    /// </summary>
-    public int MaxAllowedNodeBatchSize { get; set; } = 50;
-
-    /// <summary>
-    /// Specifies whether a plural <c>Query.nodes</c> field shall be added to the schema.
-    /// </summary>
-    public bool AddNodesField { get; set; } = true;
-
-    /// <summary>
-    /// Specifies whether the <c>Query.node</c> field should be annotated
-    /// with the composite schema <c>@lookup</c> directive.
-    /// </summary>
-    /// <remarks>
-    /// This is necessary if you want to enable the global object identification
-    /// integration in the Fusion gateway.
-    /// </remarks>
-    public bool MarkNodeFieldAsLookup { get; set; }
-}
